[Synthetics] Remove excess tests and adjust deployment agnostic tests for additional coverage (#206508)
## Summary

Remove overlapping Synthetics tests from the `api_integration` folder.

Tests are kept in the original directory if one of the following is true

A.) They haven't yet been moved to the deployment agnostic
B.) They create custom users and roles (this provides some coverage for
real stateful privileges, which the deployment agnostic directory mocks
with api keys),
C.) They deal directly with api key privileges
D.) They would fail entirely in deployment agnostic (for example, params
which fails in serverless, both MKI and local serverless).

Additionally, some files in the `deployment_agnostic` folder were split
in order to provide some level of coverage for public locations. Files
which contain coverage for public locations are skipped in cloud and
MKI, as they would fail there due to not having the mock location
available.

---------

(cherry picked from commit 93c2bd517516d93de429c74241498c0d90cf5c93)
