+++ x-pack/test/api_integration/apis/synthetics/index.ts
@@ -17,25 +17,13 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
     });
 
     loadTestFile(require.resolve('./synthetics_api_security'));
-    loadTestFile(require.resolve('./edit_monitor_public_api'));
-    loadTestFile(require.resolve('./add_monitor_public_api'));
     loadTestFile(require.resolve('./synthetics_enablement'));
-    loadTestFile(require.resolve('./get_filters'));
-    loadTestFile(require.resolve('./enable_default_alerting'));
-    loadTestFile(require.resolve('./get_monitor'));
     loadTestFile(require.resolve('./add_monitor'));
     loadTestFile(require.resolve('./add_monitor_project'));
-    loadTestFile(require.resolve('./get_monitor_project'));
     loadTestFile(require.resolve('./add_monitor_private_location'));
     loadTestFile(require.resolve('./edit_monitor'));
-    loadTestFile(require.resolve('./delete_monitor'));
-    loadTestFile(require.resolve('./delete_monitor_project'));
     loadTestFile(require.resolve('./sync_global_params'));
     loadTestFile(require.resolve('./add_edit_params'));
-    loadTestFile(require.resolve('./add_monitor_project_private_location'));
-    loadTestFile(require.resolve('./inspect_monitor'));
-    loadTestFile(require.resolve('./test_now_monitor'));
-    loadTestFile(require.resolve('./suggestions'));
     loadTestFile(require.resolve('./private_location_apis'));
     loadTestFile(require.resolve('./list_monitors'));
   });
