+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/synthetics/create_monitor.ts
@@ -6,19 +6,16 @@
  */
 import expect from '@kbn/expect';
 import { RoleCredentials, SamlAuthProviderType } from '@kbn/ftr-common-functional-services';
-import epct from 'expect';
 import moment from 'moment/moment';
 import { v4 as uuidv4 } from 'uuid';
 import { omit, omitBy } from 'lodash';
 import {
   ConfigKey,
-  MonitorTypeEnum,
   HTTPFields,
   PrivateLocation,
 } from '@kbn/synthetics-plugin/common/runtime_types';
 import { formatKibanaNamespace } from '@kbn/synthetics-plugin/common/formatters';
 import { SYNTHETICS_API_URLS } from '@kbn/synthetics-plugin/common/constants';
-import { DEFAULT_FIELDS } from '@kbn/synthetics-plugin/common/constants/monitor_defaults';
 import {
   removeMonitorEmptyValues,
   transformPublicKeys,
@@ -73,6 +70,7 @@ export const omitMonitorKeys = (monitor: any) => {
 
 export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
   describe('AddNewMonitorsUI', function () {
+    this.tags(['skipCloud', 'skipMKI']);
     const supertestAPI = getService('supertestWithoutAuth');
     const samlAuth = getService('samlAuth');
     const kibanaServer = getService('kibanaServer');
@@ -118,108 +116,6 @@ export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
       expect(apiResponse).eql(omitMonitorKeys(newMonitor));
     });
 
-    it('returns bad request if payload is invalid for HTTP monitor', async () => {
-      // Delete a required property to make payload invalid
-      const newMonitor = { ...httpMonitorJson, 'check.request.headers': null };
-      await addMonitorAPI(newMonitor, 400);
-    });
-
-    it('returns bad request if monitor type is invalid', async () => {
-      const newMonitor = { ...httpMonitorJson, type: 'invalid-data-steam' };
-
-      const apiResponse = await addMonitorAPI(newMonitor, 400);
-
-      expect(apiResponse.message).eql('Invalid value "invalid-data-steam" supplied to "type"');
-    });
-
-    it('can create valid monitors without all defaults', async () => {
-      // Delete a required property to make payload invalid
-      const newMonitor = {
-        name: 'Sample name',
-        type: 'http',
-        urls: 'https://elastic.co',
-        locations: [privateLocation],
-      };
-
-      const { body: apiResponse } = await addMonitorAPI(newMonitor);
-
-      expect(apiResponse).eql(
-        omitMonitorKeys({
-          ...DEFAULT_FIELDS[MonitorTypeEnum.HTTP],
-          ...newMonitor,
-        })
-      );
-    });
-
-    it('can disable retries', async () => {
-      const maxAttempts = 1;
-      const newMonitor = {
-        max_attempts: maxAttempts,
-        urls: 'https://elastic.co',
-        name: `Sample name ${uuidv4()}`,
-        type: 'http',
-        locations: [privateLocation],
-      };
-
-      const { body: apiResponse } = await addMonitorAPI(newMonitor);
-
-      epct(apiResponse).toEqual(epct.objectContaining({ retest_on_failure: false }));
-    });
-
-    it('can enable retries with max attempts', async () => {
-      const maxAttempts = 2;
-      const newMonitor = {
-        max_attempts: maxAttempts,
-        urls: 'https://elastic.co',
-        name: `Sample name ${uuidv4()}`,
-        type: 'http',
-        locations: [privateLocation],
-      };
-
-      const { body: apiResponse } = await addMonitorAPI(newMonitor);
-
-      epct(apiResponse).toEqual(epct.objectContaining({ retest_on_failure: true }));
-    });
-
-    it('can enable retries', async () => {
-      const newMonitor = {
-        retest_on_failure: false,
-        urls: 'https://elastic.co',
-        name: `Sample name ${uuidv4()}`,
-        type: 'http',
-        locations: [privateLocation],
-      };
-
-      const { body: apiResponse } = await addMonitorAPI(newMonitor);
-
-      epct(apiResponse).toEqual(epct.objectContaining({ retest_on_failure: false }));
-    });
-
-    it('cannot create a invalid monitor without a monitor type', async () => {
-      // Delete a required property to make payload invalid
-      const newMonitor = {
-        name: 'Sample name',
-        url: 'https://elastic.co',
-        locations: [privateLocation],
-      };
-      await addMonitorAPI(newMonitor, 400);
-    });
-
-    it('omits unknown keys', async () => {
-      // Delete a required property to make payload invalid
-      const newMonitor = {
-        name: 'Sample name',
-        url: 'https://elastic.co',
-        unknownKey: 'unknownValue',
-        type: 'http',
-        locations: [privateLocation],
-      };
-      const apiResponse = await addMonitorAPI(newMonitor, 400);
-      expect(apiResponse.message).not.to.have.keys(
-        'Invalid monitor key(s) for http type:  unknownKey","attributes":{"details":"Invalid monitor key(s) for http type:  unknownKey'
-      );
-    });
-
     it('sets namespace to Kibana space when not set to a custom namespace', async () => {
       const SPACE_ID = `test-space-${uuidv4()}`;
       const SPACE_NAME = `test-space-name ${uuidv4()}`;
@@ -243,59 +139,6 @@ export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
           .expect(200);
         monitorId = apiResponse.body.id;
         expect(apiResponse.body[ConfigKey.NAMESPACE]).eql(EXPECTED_NAMESPACE);
-      } finally {
-        await deleteMonitor(monitorId, 200, SPACE_ID);
-      }
-    });
-
-    it('preserves the passed namespace when preserve_namespace is passed', async () => {
-      const SPACE_ID = `test-space-${uuidv4()}`;
-      const SPACE_NAME = `test-space-name ${uuidv4()}`;
-      privateLocation = await privateLocationsService.addTestPrivateLocation(SPACE_ID);
-      const monitor = {
-        ...httpMonitorJson,
-        [ConfigKey.NAMESPACE]: 'default',
-        locations: [privateLocation],
-      };
-      let monitorId = '';
-      await kibanaServer.spaces.create({ id: SPACE_ID, name: SPACE_NAME });
-
-      try {
-        const apiResponse = await supertestAPI
-          .post(`/s/${SPACE_ID}${SYNTHETICS_API_URLS.SYNTHETICS_MONITORS}`)
-          .query({ preserve_namespace: true })
-          .set(editorRoleAuthc.apiKeyHeader)
-          .set(samlAuth.getInternalRequestHeader())
-          .send(monitor)
-          .expect(200);
-        monitorId = apiResponse.body.id;
-        expect(apiResponse.body[ConfigKey.NAMESPACE]).eql('default');
-      } finally {
-        await deleteMonitor(monitorId, 200, SPACE_ID);
-      }
-    });
-
-    it('sets namespace to custom namespace when set', async () => {
-      const SPACE_ID = `test-space-${uuidv4()}`;
-      const SPACE_NAME = `test-space-name ${uuidv4()}`;
-      privateLocation = await privateLocationsService.addTestPrivateLocation(SPACE_ID);
-      const monitor = {
-        ...httpMonitorJson,
-        locations: [privateLocation],
-      };
-      let monitorId = '';
-
-      try {
-        await kibanaServer.spaces.create({ id: SPACE_ID, name: SPACE_NAME });
-
-        const apiResponse = await supertestAPI
-          .post(`/s/${SPACE_ID}${SYNTHETICS_API_URLS.SYNTHETICS_MONITORS}`)
-          .set(editorRoleAuthc.apiKeyHeader)
-          .set(samlAuth.getInternalRequestHeader())
-          .send(monitor)
-          .expect(200);
-        monitorId = apiResponse.body.id;
-        expect(apiResponse.body[ConfigKey.NAMESPACE]).eql(monitor[ConfigKey.NAMESPACE]);
       } finally {
         await deleteMonitor(monitorId, 200, SPACE_ID);
       }
