+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/synthetics/create_monitor_public_api_private_location.ts
@@ -5,26 +5,33 @@
  * 2.0.
  */
 import expect from '@kbn/expect';
+import rawExpect from 'expect';
 import { v4 as uuidv4 } from 'uuid';
-
+import { RoleCredentials } from '@kbn/ftr-common-functional-services';
+import { PrivateLocation } from '@kbn/synthetics-plugin/common/runtime_types';
 import { DEFAULT_FIELDS } from '@kbn/synthetics-plugin/common/constants/monitor_defaults';
 import { LOCATION_REQUIRED_ERROR } from '@kbn/synthetics-plugin/server/routes/monitor_cruds/monitor_validation';
-import { FtrProviderContext } from '../../ftr_provider_context';
-import { addMonitorAPIHelper, omitMonitorKeys } from './add_monitor';
+import { DeploymentAgnosticFtrProviderContext } from '../../../ftr_provider_context';
+import { addMonitorAPIHelper, omitMonitorKeys } from './create_monitor';
+import { PrivateLocationTestService } from '../../../services/synthetics_private_location';
 
-export default function ({ getService }: FtrProviderContext) {
-  describe('AddNewMonitorsPublicAPI', function () {
-    this.tags('skipCloud');
-
-    const supertestAPI = getService('supertest');
+export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
+  describe('AddNewMonitorsPublicAPI - Private locations', function () {
+    const supertestAPI = getService('supertestWithoutAuth');
     const kibanaServer = getService('kibanaServer');
+    const samlAuth = getService('samlAuth');
+    let editorUser: RoleCredentials;
+    let privateLocation: PrivateLocation;
+    const privateLocationTestService = new PrivateLocationTestService(getService);
 
     async function addMonitorAPI(monitor: any, statusCode: number = 200) {
-      return await addMonitorAPIHelper(supertestAPI, monitor, statusCode);
+      return await addMonitorAPIHelper(supertestAPI, monitor, statusCode, editorUser, samlAuth);
     }
 
     before(async () => {
       await kibanaServer.savedObjects.cleanStandardList();
+      editorUser = await samlAuth.createM2mApiKeyWithRoleScope('editor');
+      privateLocation = await privateLocationTestService.addTestPrivateLocation();
     });
 
     after(async () => {
@@ -43,8 +50,8 @@ export default function ({ getService }: FtrProviderContext) {
 
     it('return error if invalid location specified', async () => {
       const { message } = await addMonitorAPI({ type: 'http', locations: ['mars'] }, 400);
-      expect(message).eql(
-        "Invalid locations specified. Elastic managed Location(s) 'mars' not found. Available locations are 'dev|dev2'"
+      rawExpect(message).toContain(
+        "Invalid locations specified. Elastic managed Location(s) 'mars' not found."
       );
     });
 
@@ -67,20 +74,8 @@ export default function ({ getService }: FtrProviderContext) {
         },
         400
       );
-      expect(result.message).eql(
-        "Invalid locations specified. Elastic managed Location(s) 'mars' not found. Available locations are 'dev|dev2' Private Location(s) 'moon' not found. No private location available to use."
-      );
+      rawExpect(result.message).toContain("Private Location(s) 'moon' not found.");
     });
-
-    const localLoc = {
-      id: 'dev',
-      label: 'Dev Service',
-      geo: {
-        lat: 0,
-        lon: 0,
-      },
-      isServiceManaged: true,
-    };
 
     it('return error for origin project', async () => {
       const { message } = await addMonitorAPI(
@@ -101,7 +96,8 @@ export default function ({ getService }: FtrProviderContext) {
         const { message, attributes } = await addMonitorAPI(
           {
             type: 'http',
-            locations: ['dev'],
+            locations: [],
+            private_locations: [privateLocation.id],
           },
           400
         );
@@ -117,7 +113,7 @@ export default function ({ getService }: FtrProviderContext) {
       it('base http monitor', async () => {
         const monitor = {
           type: 'http',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           url: 'https://www.google.com',
         };
         const { body: result } = await addMonitorAPI(monitor);
@@ -126,7 +122,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
             name: 'https://www.google.com',
           })
         );
@@ -136,7 +132,7 @@ export default function ({ getService }: FtrProviderContext) {
         const name = `test name ${uuidv4()}`;
         const monitor = {
           type: 'http',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           url: 'https://www.google.com',
           name,
           retest_on_failure: true,
@@ -147,7 +143,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
             name,
             retest_on_failure: true,
           })
@@ -158,7 +154,7 @@ export default function ({ getService }: FtrProviderContext) {
         const name = `test name ${uuidv4()}`;
         const monitor = {
           type: 'http',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           url: 'https://www.google.com',
           name,
           retest_on_failure: false,
@@ -169,7 +165,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
             name,
             max_attempts: 1,
             retest_on_failure: undefined, // this key is not part of the SO and should not be defined
@@ -184,7 +180,7 @@ export default function ({ getService }: FtrProviderContext) {
       it('base tcp monitor', async () => {
         const monitor = {
           type: 'tcp',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           host: 'https://www.google.com/',
         };
         const { body: result } = await addMonitorAPI(monitor);
@@ -193,7 +189,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
             name: 'https://www.google.com/',
           })
         );
@@ -206,7 +202,7 @@ export default function ({ getService }: FtrProviderContext) {
       it('base icmp monitor', async () => {
         const monitor = {
           type: 'icmp',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           host: 'https://8.8.8.8',
         };
         const { body: result } = await addMonitorAPI(monitor);
@@ -215,7 +211,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
             name: 'https://8.8.8.8',
           })
         );
@@ -228,7 +224,7 @@ export default function ({ getService }: FtrProviderContext) {
       it('empty browser monitor', async () => {
         const monitor = {
           type: 'browser',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           name: 'simple journey',
         };
         const result = await addMonitorAPI(monitor, 400);
@@ -247,7 +243,7 @@ export default function ({ getService }: FtrProviderContext) {
       it('base browser monitor', async () => {
         const monitor = {
           type: 'browser',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           name: 'simple journey',
           'source.inline.script': 'step("simple journey", async () => {});',
         };
@@ -257,7 +253,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
           })
         );
       });
@@ -265,7 +261,7 @@ export default function ({ getService }: FtrProviderContext) {
       it('base browser monitor with inline_script', async () => {
         const monitor = {
           type: 'browser',
-          locations: ['dev'],
+          private_locations: [privateLocation.id],
           name: 'simple journey inline_script',
           inline_script: 'step("simple journey", async () => {});',
         };
@@ -275,7 +271,7 @@ export default function ({ getService }: FtrProviderContext) {
           omitMonitorKeys({
             ...defaultFields,
             ...monitor,
-            locations: [localLoc],
+            locations: [privateLocation],
           })
         );
       });
