+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/synthetics/helpers/location.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const LOCAL_PUBLIC_LOCATION = {
+  geo: {
+    lat: 0,
+    lon: 0,
+  },
+  id: 'dev',
+  label: 'Dev Service',
+  isServiceManaged: true,
+};
